Maintenance of the error path corrected
The error path is collected only when the error is pased on to the entry procedure. When the error is displayed immediately, either because the entry procedure is unknown or because more then one buttons are about to be displayed, ther error path is empty and only ther error source is displayed in the message.